requirements.txt
google-genai
pydantic
rich
python-dotenv
camel-oasis>=0.2.5
